fixing employeeServjce + adding loggers

+++ src/main/java/tn/esprit/spring/services/EmployeServiceImpl.java
@@ -1,13 +1,12 @@
 package tn.esprit.spring.services;
 
-import java.io.Console;
 import java.util.ArrayList;
-
 import java.util.Date;
 import java.util.List;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -23,10 +22,16 @@ import tn.esprit.spring.repository.DepartementRepository;
 import tn.esprit.spring.repository.EmployeRepository;
 import tn.esprit.spring.repository.TimesheetRepository;
 
+
+import java.util.Optional;
+
+
+
 @Service
 public class EmployeServiceImpl implements IEmployeService {
 
- Logger logger = LoggerFactory.getLogger(EmployeServiceImpl.class);
+	private static final Logger l = LoggerFactory.getLogger(EmployeServiceImpl.class);
+
 	@Autowired
 	EmployeRepository employeRepository;
 	@Autowired
@@ -36,60 +41,101 @@ public class EmployeServiceImpl implements IEmployeService {
 	@Autowired
 	TimesheetRepository timesheetRepository;
 
-	/*public int ajouterEmploye(Employe employe) {
-		employeRepository.save(employe);
-		return employe.getId();
-	}*/
-	
+	@Override
 	public int ajouterEmploye(Employe employe) {
 		try {
-			
-			logger.trace("je vais ajouter un employer");
-			System.out.println("anis");
 			employeRepository.save(employe);
-	
+			l.info(employe.toString());
 			return employe.getId();
+
 		} catch (Exception e) {
-			logger.error("erreur dans ajouterEmploye:"+e);
+			l.error("Add failure");
+			return 0;
+		}
+
+	}
+
+	public int mettreAjourEmailByEmployeId(String email, int employeId) {
+
+		try {
+			Optional<Employe> employeManagedEntity = employeRepository.findById(employeId);
+
+			if (employeManagedEntity.isPresent()) {
+				Employe employe = employeManagedEntity.get();
+				employe.setEmail(email);
+				employeRepository.save(employe);
+				l.info("Employe Updated: " + employe);
+				return employeId;
+
+			} else {
+
+				l.warn("Employe don't exist");
+				return 0;
+
+			}
+
+		} catch (Exception e) {
+
+			l.error(e.toString());
+			return 0;
+		}
+
+	}
+
+	@Transactional
+	public int affecterEmployeADepartement(int employeId, int depId) {
+		try {
+
+			Optional<Departement> depManagedEntity = deptRepoistory.findById(depId);
+			Optional<Employe> employeManagedEntity = employeRepository.findById(employeId);
+
+			if (depManagedEntity.isPresent() && employeManagedEntity.isPresent()) {
+
+				Departement department = depManagedEntity.get();
+				Employe employe = employeManagedEntity.get();
+
+				if (department.getEmployes() == null) {
+
+					List<Employe> employes = new ArrayList<>();
+					employes.add(employe);
+					department.setEmployes(employes);
+				} else {
+
+					department.getEmployes().add(employe);
+				}
+
+				deptRepoistory.save(department);
+
+				l.info("departement affected " + department);
+				l.info("employee affected " + employe);
+				return employeId ;
+
+			}
+
+			else {
+				l.warn("Cet employe ou ce departement n'existe pas");
+				return 0;
+			}
+
+		} catch (Exception e) {
+			l.error(e.toString());
 			return 0;
 		}
 	}
 
-	public void mettreAjourEmailByEmployeId(String email, int employeId) {
-		Employe employe = employeRepository.findById(employeId).get();
-		employe.setEmail(email);
-		employeRepository.save(employe);
-
-	}
-
-	@Transactional	
-	public void affecterEmployeADepartement(int employeId, int depId) {
-		Departement depManagedEntity = deptRepoistory.findById(depId).get();
-		Employe employeManagedEntity = employeRepository.findById(employeId).get();
-
-		if(depManagedEntity.getEmployes() == null){
-
-			List<Employe> employes = new ArrayList<>();
-			employes.add(employeManagedEntity);
-			depManagedEntity.setEmployes(employes);
-		}else{
-
-			depManagedEntity.getEmployes().add(employeManagedEntity);
-
-		}
-
-	}
 	@Transactional
-	public void desaffecterEmployeDuDepartement(int employeId, int depId)
-	{
-		Departement dep = deptRepoistory.findById(depId).get();
-
-		int employeNb = dep.getEmployes().size();
-		for(int index = 0; index < employeNb; index++){
-			if(dep.getEmployes().get(index).getId() == employeId){
-				dep.getEmployes().remove(index);
-				break;//a revoir
+	public void desaffecterEmployeDuDepartement(int employeId, int depId) {
+		Departement dep = deptRepoistory.findById(depId).orElse(null);
+		if (dep != null) {
+			int employeNb = dep.getEmployes().size();
+			for (int index = 0; index < employeNb; index++) {
+				if (dep.getEmployes().get(index).getId() == employeId) {
+					dep.getEmployes().remove(index);
+					break;
+				}
 			}
+		} else {
+			l.warn("Dep don't exist");
 		}
 	}
 
@@ -99,47 +145,68 @@ public class EmployeServiceImpl implements IEmployeService {
 	}
 
 	public void affecterContratAEmploye(int contratId, int employeId) {
-		Contrat contratManagedEntity = contratRepoistory.findById(contratId).get();
-		Employe employeManagedEntity = employeRepository.findById(employeId).get();
+		Contrat contratManagedEntity = contratRepoistory.findById(contratId).orElse(null);
+		Employe employeManagedEntity = employeRepository.findById(employeId).orElse(null);
 
-		contratManagedEntity.setEmploye(employeManagedEntity);
-		contratRepoistory.save(contratManagedEntity);
-		
+		if ((contratManagedEntity != null) && (employeManagedEntity != null)) {
+			contratManagedEntity.setEmploye(employeManagedEntity);
+			contratRepoistory.save(contratManagedEntity);
+			l.info("Contract affected");
+
+		} else {
+			l.warn("Employe or Contract not found");
+		}
+
 	}
 
 	public String getEmployePrenomById(int employeId) {
-		Employe employeManagedEntity = employeRepository.findById(employeId).get();
-		return employeManagedEntity.getPrenom();
-	}
-	public void deleteEmployeById(int employeId)
-	{
-		Employe employe = employeRepository.findById(employeId).get();
+		Employe employeManagedEntity = employeRepository.findById(employeId).orElse(null);
+		if (employeManagedEntity != null) {
+			return employeManagedEntity.getPrenom();
 
-		//Desaffecter l'employe de tous les departements
-		//c'est le bout master qui permet de mettre a jour
-		//la table d'association
-		for(Departement dep : employe.getDepartements()){
-			dep.getEmployes().remove(employe);
+		} else {
+			l.warn("Employe don't exist");
 		}
+		return null;
+	}
 
-		employeRepository.delete(employe);
+	public int deleteEmployeById(int employeId) {
+		Employe employe = employeRepository.findById(employeId).orElse(null);
+
+		if (employe != null) {
+			for (Departement dep : employe.getDepartements()) {
+				dep.getEmployes().remove(employe);
+			}
+
+			employeRepository.delete(employe);
+			l.info("employe deleted");
+			return 1;
+		} else {
+			l.warn("Employe dosen't exist");
+			return 0;
+		}
 	}
 
 	public void deleteContratById(int contratId) {
-		Contrat contratManagedEntity = contratRepoistory.findById(contratId).get();
-		contratRepoistory.delete(contratManagedEntity);
+		Contrat contratManagedEntity = contratRepoistory.findById(contratId).orElse(null);
+		if (contratManagedEntity != null) {
+			contratRepoistory.delete(contratManagedEntity);
+			l.info("Contract Deleted");
+		} else {
+			l.warn("Contract not found");
+		}
 
 	}
 
 	public int getNombreEmployeJPQL() {
 		return employeRepository.countemp();
 	}
-	
+
 	public List<String> getAllEmployeNamesJPQL() {
 		return employeRepository.employeNames();
 
 	}
-	
+
 	public List<Employe> getAllEmployeByEntreprise(Entreprise entreprise) {
 		return employeRepository.getAllEmployeByEntreprisec(entreprise);
 	}
@@ -148,10 +215,11 @@ public class EmployeServiceImpl implements IEmployeService {
 		employeRepository.mettreAjourEmailByEmployeIdJPQL(email, employeId);
 
 	}
+
 	public void deleteAllContratJPQL() {
-         employeRepository.deleteAllContratJPQL();
+		employeRepository.deleteAllContratJPQL();
 	}
-	
+
 	public float getSalaireByEmployeIdJPQL(int employeId) {
 		return employeRepository.getSalaireByEmployeIdJPQL(employeId);
 	}
@@ -159,14 +227,17 @@ public class EmployeServiceImpl implements IEmployeService {
 	public Double getSalaireMoyenByDepartementId(int departementId) {
 		return employeRepository.getSalaireMoyenByDepartementId(departementId);
 	}
-	
+
 	public List<Timesheet> getTimesheetsByMissionAndDate(Employe employe, Mission mission, Date dateDebut,
-			Date dateFin) {
+														 Date dateFin) {
 		return timesheetRepository.getTimesheetsByMissionAndDate(employe, mission, dateDebut, dateFin);
 	}
 
 	public List<Employe> getAllEmployes() {
-				return (List<Employe>) employeRepository.findAll();
+		return (List<Employe>) employeRepository.findAll();
 	}
 
+
+
+	
 }
